+++ src/gmail_automation/__init__.py
@@ -1,0 +1,1 @@
+
